Some small fixes and improvements
* Fixes linking of LLVM IR modules
* Maps some more LLVM optimization flags to instruction decorations (nsw, nuw, exact, dereferenceable)
* Removes some unused code
* Fixes compilation errors for vc4asm validator under GCC 9
* Extends tests for different front-ends
* Adds flag behavior as constants to opcodes (to be used later...)
* Reactivates SPIR-V build and test in CircleCI

+++ testing/test_linking_1.cl
@@ -1,5 +1,5 @@
 
-int16 get_value(const __global int* in)
+int get_value(const __global int* in)
 {
     return *in;
 }
